Decouple logic into separate files
